Rename class & add test for opening help fragment
